+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog (nionui-tool)
 0.4.9 (UNRELEASED)
 ------------------
 - Add method to set window file path.
+- Add method to query keyboard modifiers directly.
 
 0.4.8 (2020-12-28)
 ------------------